+++ src/cli/commands.js
@@ -1,9 +1,22 @@
-/* jshint maxlen: false */
 'use strict';
 
 module.exports = {
   usage: 'Usage: $0 <command> [options]',
   commands: {
+    clean: {
+      cmd: 'clean <dirPath>',
+      example: { cmd: '$0 clean components --yes' },
+      description: `Remove the node_modules directory from each component's subfolder`,
+      options: {
+        yes: {
+          boolean: true,
+          description: 'Skip all confirmation prompts',
+          default: false
+        }
+      },
+      usage: 'Usage: $0 clean <dirPath>'
+    },
+
     dev: {
       cmd: 'dev <dirPath> [port] [baseUrl]',
       example: {
@@ -11,7 +24,7 @@ module.exports = {
           '$0 dev ../all-components 3001 127.0.0.1:3001 --fallbackRegistryUrl=http://anotherhost:anotherport/'
       },
       description:
-        'Runs a local oc test registry in order to develop and test components',
+        'Run a local oc test registry in order to develop and test components',
       options: {
         fallbackRegistryUrl: {
           description:
@@ -20,7 +33,7 @@ module.exports = {
         hotReloading: {
           boolean: true,
           description:
-            'Enables hot reloading. Note: when hot reloading is set to true, each request to the component will make the registry to create a new instance for the javascript closures to be loaded, while when false the instance will be recycled between components executions',
+            'Enable hot reloading. Note: when hot reloading is set to true, each request to the component will make the registry to create a new instance for the javascript closures to be loaded, while when false the instance will be recycled between components executions',
           default: true
         },
         verbose: {
@@ -44,17 +57,17 @@ module.exports = {
           default: ''
         }
       },
-      usage: 'Usage: $0 dev <dirName> [port] [baseUrl] [options]'
+      usage: 'Usage: $0 dev <dirPath> [port] [baseUrl] [options]'
     },
 
     init: {
-      cmd: 'init <componentName> [templateType]',
+      cmd: 'init <componentPath> [templateType]',
       example: {
-        cmd: '$0 init test-component jade'
+        cmd: '$0 init test-component oc-template-es6'
       },
       description:
-        'Creates a new empty component of a specific template type in the current folder [templateType default: oc-template-handlebars]',
-      usage: 'Usage: $0 init <componentName> [templateType]'
+        'Create a new empty component of a specific template type in the current folder or in directory indicated by (relative) componentPath [templateType default: oc-template-es6]',
+      usage: 'Usage: $0 init <componentPath> [templateType]'
     },
 
     mock: {
@@ -62,7 +75,7 @@ module.exports = {
       example: {
         cmd: '$0 mock plugin hash "always-returned-value"',
         description:
-          'Creates static mock for a "hash" plugin which always returns "always-returned-value" value'
+          'Create static mock for a "hash" plugin which always returns "always-returned-value" value'
       },
       description:
         'Allows to mock configuration in order to facilitate local development',
@@ -75,7 +88,7 @@ module.exports = {
         cmd:
           '$0 preview "http://localhost:3000/my-new-component/1.0.0/?param1=hello&name=Arthur"'
       },
-      description: 'Runs a test page consuming a component',
+      description: 'Run a test page consuming a component',
       usage: 'Usage: $0 preview <componentHref>'
     },
 
@@ -89,9 +102,15 @@ module.exports = {
           boolean: true,
           description: 'Create zipped file',
           default: false
+        },
+        useComponentDependencies: {
+          boolean: true,
+          description:
+            'Reuse already installed dependencies to save packaging time and network bandwidth',
+          default: false
         }
       },
-      description: 'Creates the packaged component ready to be published',
+      description: 'Create the packaged component ready to be published',
       usage: 'Usage: $0 package <componentPath>'
     },
 
@@ -116,21 +135,21 @@ module.exports = {
 
     registry: {
       cmd: 'registry <command>',
-      description: 'Manages oc registries in the current project',
+      description: 'Manage oc registries in the current project',
       commands: {
         add: {
           cmd: 'add <registryUrl>',
           example: {
             cmd: '$0 registry add http://my-registry.in.my.domain/'
           },
-          description: 'Adds oc registries to the current project',
+          description: 'Add oc registries to the current project',
           usage: 'Usage: $0 registry add <registryUrl>'
         },
         ls: {
           example: {
             cmd: '$0 registry ls'
           },
-          description: 'Shows oc registries added to the current project',
+          description: 'Show oc registries added to the current project',
           usage: 'Usage: $0 registry ls'
         },
         remove: {
@@ -138,7 +157,7 @@ module.exports = {
           example: {
             cmd: '$0 registry remove http://my-registry.in.my.domain/'
           },
-          description: 'Removes oc registries from the current project',
+          description: 'Remove oc registries from the current project',
           usage: 'Usage: $0 registry remove <registryUrl>'
         }
       },