+++ src/cli/domain/local.js
@@ -5,15 +5,18 @@ const path = require('path');
 const targz = require('targz');
 const _ = require('lodash');
 
+const clean = require('./clean');
 const getComponentsByDir = require('./get-components-by-dir');
-const packageComponents = require('./package-components');
-const mock = require('./mock');
-const validator = require('../../registry/domain/validators');
 const initTemplate = require('./init-template');
+const isTemplateLegacy = require('../../utils/is-template-legacy');
+const mock = require('./mock');
+const packageComponents = require('./package-components');
 const strings = require('../../resources');
+const validator = require('../../registry/domain/validators');
 
 module.exports = function() {
   return _.extend(this, {
+    clean,
     cleanup: function(compressedPackagePath, callback) {
       return fs.unlink(compressedPackagePath, callback);
     },
@@ -25,7 +28,7 @@ module.exports = function() {
           tar: {
             map: function(file) {
               return _.extend(file, {
-                name: '_package/' + file.name
+                name: `_package/${file.name}`
               });
             }
           }
@@ -41,7 +44,7 @@ module.exports = function() {
       }
 
       // LEGACY TEMPLATES WARNING
-      if (validator.validateTemplateType(templateType)) {
+      if (isTemplateLegacy(templateType)) {
         const legacyName = templateType;
         templateType = legacyName.replace(
           legacyName,