+++ src/registry/routes/component-preview.js
@@ -2,8 +2,9 @@
 
 const _ = require('lodash');
 
-const urlBuilder = require('../domain/url-builder');
 const getComponentFallback = require('./helpers/get-component-fallback');
+const previewView = require('../views/preview');
+const urlBuilder = require('../domain/url-builder');
 
 function componentPreview(err, req, res, component, templates) {
   if (err) {
@@ -12,24 +13,27 @@ function componentPreview(err, req, res, component, templates) {
     return res.status(404).json(err);
   }
 
-  let liveReload = null;
-  if (res.conf.env.name === 'local') {
-    const liveReloadPort = res.conf.port + 1;
-    liveReload = `<script src="http://localhost:${liveReloadPort}/livereload.js?snipver=1"></script>`;
+  let liveReload = '';
+  if (res.conf.liveReloadPort) {
+    liveReload = `<script src="http://localhost:${
+      res.conf.liveReloadPort
+    }/livereload.js?snipver=1"></script>`;
   }
 
   const isHtmlRequest =
     !!req.headers.accept && req.headers.accept.indexOf('text/html') >= 0;
 
   if (isHtmlRequest && !!res.conf.discovery) {
-    return res.render('component-preview', {
-      component: component,
-      dependencies: _.keys(component.dependencies),
-      href: res.conf.baseUrl,
-      liveReload,
-      qs: urlBuilder.queryString(req.query),
-      templates
-    });
+    return res.send(
+      previewView({
+        component,
+        dependencies: _.keys(component.dependencies),
+        href: res.conf.baseUrl,
+        liveReload,
+        qs: urlBuilder.queryString(req.query),
+        templates
+      })
+    );
   } else {
     res.status(200).json(
       _.extend(component, {