+++ src/registry/routes/index.js
@@ -7,17 +7,20 @@ const _ = require('lodash');
 const dateStringified = require('../../utils/date-stringify');
 const getComponentsHistory = require('./helpers/get-components-history');
 const getAvailableDependencies = require('./helpers/get-available-dependencies');
+const indexView = require('../views');
 const packageInfo = require('../../../package.json');
 const urlBuilder = require('../domain/url-builder');
 
-function getParsedAuthor(author = {}) {
+const getParsedAuthor = author => {
+  author = author || {};
   return _.isString(author) ? parseAuthor(author) : author;
-}
+};
 
-function mapComponentDetails(component) {
-  component.author = getParsedAuthor(component.author);
-  return component;
-}
+const mapComponentDetails = component =>
+  _.extend(component, { author: getParsedAuthor(component.author) });
+
+const isHtmlRequest = headers =>
+  !!headers.accept && headers.accept.indexOf('text/html') >= 0;
 
 module.exports = function(repository) {
   return function(req, res, next) {
@@ -27,15 +30,13 @@ module.exports = function(repository) {
         return res.status(404).json({ error: res.errorDetails });
       }
 
-      const isHtmlRequest =
-          !!req.headers.accept && req.headers.accept.indexOf('text/html') >= 0,
-        baseResponse = {
-          href: res.conf.baseUrl,
-          ocVersion: packageInfo.version,
-          type: res.conf.local ? 'oc-registry-local' : 'oc-registry'
-        };
+      const baseResponse = {
+        href: res.conf.baseUrl,
+        ocVersion: packageInfo.version,
+        type: res.conf.local ? 'oc-registry-local' : 'oc-registry'
+      };
 
-      if (isHtmlRequest && !!res.conf.discovery) {
+      if (isHtmlRequest(req.headers) && !!res.conf.discovery) {
         let componentsInfo = [],
           componentsReleases = 0;
         const stateCounts = {};
@@ -44,9 +45,7 @@ module.exports = function(repository) {
           components,
           (component, callback) =>
             repository.getComponent(component, (err, result) => {
-              if (err) {
-                return callback(err);
-              }
+              if (err) return callback(err);
 
               if (result.oc && result.oc.date) {
                 result.oc.stringifiedDate = dateStringified(
@@ -59,44 +58,41 @@ module.exports = function(repository) {
               callback();
             }),
           err => {
-            if (err) {
-              return next(err);
-            }
+            if (err) return next(err);
 
             componentsInfo = _.sortBy(componentsInfo, 'name');
-
             repository.getComponentsDetails((err, details) => {
-              res.render(
-                'index',
-                _.extend(baseResponse, {
-                  availableDependencies: getAvailableDependencies(
-                    res.conf.dependencies
-                  ),
-                  availablePlugins: res.conf.plugins,
-                  components: componentsInfo,
-                  componentsReleases,
-                  componentsList: _.map(componentsInfo, component => {
-                    const state =
-                      !!component.oc && !!component.oc.state
-                        ? component.oc.state
-                        : '';
+              if (err) console.log(err);
+              res.send(
+                indexView(
+                  _.extend(baseResponse, {
+                    availableDependencies: getAvailableDependencies(
+                      res.conf.dependencies
+                    ),
+                    availablePlugins: res.conf.plugins,
+                    components: componentsInfo,
+                    componentsReleases,
+                    componentsList: _.map(componentsInfo, component => {
+                      const state = _.get(component, 'oc.state', '');
+                      if (state) {
+                        stateCounts[state] = stateCounts[state] || 0;
+                        stateCounts[state] += 1;
+                      }
 
-                    if (state) {
-                      stateCounts[state] = stateCounts[state] || 0;
-                      stateCounts[state] += 1;
-                    }
-
-                    return {
-                      name: component.name,
-                      author: component.author,
-                      state: state
-                    };
-                  }),
-                  componentsHistory:
-                    !res.conf.local && getComponentsHistory(details),
-                  q: req.query.q || '',
-                  stateCounts
-                })
+                      return {
+                        name: component.name,
+                        author: component.author,
+                        state
+                      };
+                    }),
+                    componentsHistory:
+                      !res.conf.local && getComponentsHistory(details),
+                    q: req.query.q || '',
+                    stateCounts,
+                    templates: repository.getTemplatesInfo(),
+                    title: 'OpenComponents Registry'
+                  })
+                )
               );
             });
           }