+++ test/unit/registry-domain-validator.js
@@ -765,33 +765,6 @@ describe('registry : domain : validator', () => {
     });
   });
 
-  describe('when validating template type for new candidate', () => {
-    const validate = function(a) {
-      return validator.validateTemplateType(a);
-    };
-
-    describe('when type is not handlebars or jade', () => {
-      const type = 'othertype';
-      it('should not be valid', () => {
-        expect(validate(type)).to.be.false;
-      });
-    });
-
-    describe('when type is handlebars', () => {
-      const type = 'handlebars';
-      it('should be valid', () => {
-        expect(validate(type)).to.be.true;
-      });
-    });
-
-    describe('when type is jade', () => {
-      const type = 'jade';
-      it('should be valid', () => {
-        expect(validate(type)).to.be.true;
-      });
-    });
-  });
-
   describe('when validating component version for new candidate', () => {
     const existingVersions = ['1.0.0', '1.0.1', '2.0.0', '2.1.0'],
       isValid = function(a, b) {
@@ -1060,7 +1033,7 @@ describe('registry : domain : validator', () => {
       });
 
       it('should not return an error', () => {
-        expect(result.error).to.be.empty;
+        expect(result.error).to.be.undefined;
       });
     });
 
@@ -1072,7 +1045,7 @@ describe('registry : domain : validator', () => {
       });
 
       it('should not return an error', () => {
-        expect(result.error).to.be.empty;
+        expect(result.error).to.be.undefined;
       });
     });
   });
@@ -1135,7 +1108,7 @@ describe('registry : domain : validator', () => {
       });
 
       it('should not return an error', () => {
-        expect(result.error).to.be.empty;
+        expect(result.error).to.be.undefined;
       });
     });
 
@@ -1147,7 +1120,7 @@ describe('registry : domain : validator', () => {
       });
 
       it('should not return an error', () => {
-        expect(result.error).to.be.empty;
+        expect(result.error).to.be.undefined;
       });
     });
   });